refactor: rename translatable module to design_creator
