+++ elasticai/creator/nn/fixed_point/conv1d/layer.py
@@ -7,7 +7,7 @@ from elasticai.creator.nn.fixed_point._math_operations import MathOperations
 from elasticai.creator.nn.fixed_point._two_complement_fixed_point_config import (
     FixedPointConfig,
 )
-from elasticai.creator.vhdl.translatable import DesignCreator
+from elasticai.creator.vhdl.design_creator import DesignCreator
 
 from .design import Conv1d as Conv1dDesign
