+++ elasticai/creator/nn/fixed_point/hard_sigmoid/layer.py
@@ -2,7 +2,7 @@ from elasticai.creator.base_modules.hard_sigmoid import HardSigmoid as HardSigmo
 from elasticai.creator.nn.fixed_point._two_complement_fixed_point_config import (
     FixedPointConfig,
 )
-from elasticai.creator.vhdl.translatable import DesignCreator
+from elasticai.creator.vhdl.design_creator import DesignCreator
 
 from .design import HardSigmoid as HardSigmoidDesign
