+++ elasticai/creator/nn/sequential/layer.py
@@ -4,7 +4,7 @@ from typing import cast
 import torch
 
 from elasticai.creator.vhdl.design.design import Design
-from elasticai.creator.vhdl.translatable import DesignCreator
+from elasticai.creator.vhdl.design_creator import DesignCreator
 
 from .design import Sequential as _SequentialDesign
